+++ app/src/main/java/org/simple/clinic/settings/SettingsEffect.kt
@@ -10,9 +10,11 @@ object LoadAppVersionEffect : SettingsEffect()
 
 object CheckAppUpdateAvailable : SettingsEffect()
 
-sealed class SettingsViewEffect : SettingsEffect()
-
 object LogoutUser : SettingsEffect()
+
+object LoadDatabaseEncryptionStatus : SettingsEffect()
+
+sealed class SettingsViewEffect : SettingsEffect()
 
 object OpenLanguageSelectionScreenEffect : SettingsViewEffect()